requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Markdown==3.4.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
requests==2.30.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
